Support messageBoxRedirect in non-ide engines
