+++ docs/dictionary/message/msgChanged.lcdoc
@@ -1,0 +1,28 @@
+Name: msgChanged
+
+Type: message
+
+Syntax: msgChanged
+
+Summary:
+Sent to the <messageBoxRedirect> object whenever the <message box> is
+changed.
+
+Introduced: 9.0
+
+OS: mac, windows, linux, ios, android, html5
+
+Platforms: desktop, mobile
+
+Example:
+on msgChanged
+  set the text of field "Message" to msg
+end msgChanged
+
+Description:
+Use the msgChanged <message> to display or log the content of the
+<message box>.
+
+References: put (command), container (glossary), command (glossary),
+message (glossary), message box (glossary), property (glossary), 
+messageBoxLastObject (property), messageBoxRedirect (property)
