+++ docs/dictionary/property/messageBoxLastObject.lcdoc
@@ -1,0 +1,29 @@
+Name: messageBoxLastObject
+
+Type: property
+
+Syntax: get the messageBoxLastObject
+
+Summary:
+Get the long id of the last object to update the message box.
+
+Introduced: 9.0
+
+OS: mac, windows, linux, ios, android, html5
+
+Platforms: desktop, mobile
+
+Example:
+local tObject
+put the messageBoxLastObject into tObject
+if tObject is not empty then
+   edit the script of tObject
+end if
+
+Description:
+Use the messageBoxLastObject <property> to find the last object that
+modified the message box.
+
+References: put (command), container (glossary), command (glossary),
+message (glossary), message box (glossary), property (glossary), 
+msgChanged (message), messageBoxRedirect (property)
