+++ engine/src/parsedef.h
@@ -1562,8 +1562,8 @@ enum Properties {
     P_REV_LICENSE_LIMITS, // DEVELOPMENT only
 	P_REV_CRASH_REPORT_SETTINGS, // DEVELOPMENT only
 	P_REV_AVAILABLE_HANDLERS, // DEVELOPMENT only
-	P_REV_MESSAGE_BOX_LAST_OBJECT, // DEVELOPMENT only
-	P_REV_MESSAGE_BOX_REDIRECT, // DEVELOPMENT only
+	P_MESSAGE_BOX_LAST_OBJECT,
+	P_MESSAGE_BOX_REDIRECT,
 	P_REV_LICENSE_INFO, // DEVELOPMENT only
 
 	P_REV_RUNTIME_BEHAVIOUR,
